// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithaccongtrinh.model.ts
export class OutputCpKhaiThacCongTrinhModel {
    idcongtrinh: string;
    sohieu: string;
    chieusau: number;
    toadox: number;
    toadoy: number;
    luuluong: number;
    chedokhaithac: string;
    hathap: number;
    mucnuoctinh: number;
    ghichu: string;
    idcapphepkhaithac: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    siteid: string
}

export class InputCpKhaiThacCongTrinhModel {
    idcongtrinh: string;
    sohieu: string;
    chieusau: number;
    toadox: number;
    toadoy: number;
    luuluong: number;
    chedokhaithac: string;
    hathap: number;
    mucnuoctinh: number;
    ghichu: string;
    idcapphepkhaithac: string;
    hequychieu: string;
}